Feature/authorisation

+++ controllers/card_controller.py
@@ -6,6 +6,7 @@ from flask_jwt_extended import jwt_required, get_jwt_identity
 from init import db
 from models.card import Card, card_schema, cards_schema
 from controllers.comment_controller import comments_bp
+from utils import authorise_as_admin
 
 cards_bp = Blueprint("cards", __name__, url_prefix="/cards")
 cards_bp.register_blueprint(comments_bp)
@@ -19,7 +20,7 @@ cards_bp.register_blueprint(comments_bp)
 # /cards - GET - fetch all cards
 @cards_bp.route("/")
 def get_all_cards():
-    # fetch all cards but order them according to date in descending order (extra lil feature)
+    # fetch all cards and order them according to date in descending order
     stmt = db.select(Card).order_by(Card.date.desc())
     cards = db.session.scalars(stmt)
     return cards_schema.dump(cards)
@@ -28,15 +29,17 @@ def get_all_cards():
 @cards_bp.route("/<int:card_id>")
 def get_one_card(card_id):
     stmt = db.select(Card).filter_by(id=card_id)
+    # stmt = db.select(Card).where(Card.id==card_id)
     card = db.session.scalar(stmt)
     if card:
         return card_schema.dump(card)
     else:
         return {"error": f"Card with id {card_id} not found"}, 404
     
+
 # /cards - POST - create a new card
 @cards_bp.route("/", methods=["POST"])
-@jwt_required() # user has to be logged in to post
+@jwt_required()
 def create_card():
     # get the data from the body of the request
     body_data = card_schema.load(request.get_json())
@@ -44,28 +47,31 @@ def create_card():
     card = Card(
         title=body_data.get("title"),
         description=body_data.get("description"),
-        date=date.today(), # make sure to import library "from datetime import date"
+        date=date.today(),
         status=body_data.get("status"),
         priority=body_data.get("priority"),
         user_id=get_jwt_identity()
     )
-
     # add and commit to DB
     db.session.add(card)
     db.session.commit()
-
     # respond
     return card_schema.dump(card)
 
+
 # /cards/<id> - DELETE - delete a card
 @cards_bp.route("/<int:card_id>", methods=["DELETE"])
-@jwt_required() # user has to be logged in to delete
+@jwt_required()
 def delete_card(card_id):
     # fetch the card from the database
     stmt = db.select(Card).filter_by(id=card_id)
     card = db.session.scalar(stmt)
-    #if card
+    # if card
     if card:
+        # check whether the user is an admin or not
+        is_admin = authorise_as_admin()
+        if not is_admin and str(card.user_id) != get_jwt_identity():
+            return {"error": "User is not authorised to perform this action."}, 403
         # delete the card
         db.session.delete(card)
         db.session.commit()
@@ -74,19 +80,23 @@ def delete_card(card_id):
     else:
         # return error
         return {"error": f"Card with id {card_id} not found"}, 404
+    
 
 # /cards/<id> - PUT, PATCH - edit a card
 @cards_bp.route("/<int:card_id>", methods=["PUT", "PATCH"])
-@jwt_required() # user has to be logged in to update a card
+@jwt_required()
 def update_card(card_id):
     # get the data from the body of the request
     body_data = card_schema.load(request.get_json(), partial=True)
     # get the card from the database
     stmt = db.select(Card).filter_by(id=card_id)
     card = db.session.scalar(stmt)
-    # if card
+    # if card exists
     if card:
-        # if card exists
+        # if the user is not the owner of the card
+        if str(card.user_id) != get_jwt_identity():
+            return {"error": "You are not the owner of the card"}, 403
+        # update the fields as required
         card.title = body_data.get("title") or card.title
         card.description = body_data.get("description") or card.description
         card.status = body_data.get("status") or card.status
